AI bot/requirements.txt
pyTelegramBotAPI
google-generativeai
groq
python-dotenv
